backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
torch>=2.6.0
torchvision>=0.21.0
timm>=0.9.16
opencv-python-headless>=4.9.0.80
albumentations>=1.3.1
grad-cam>=1.4.8
Pillow>=10.2.0
numpy>=1.26.4
google-generativeai>=0.5.0
